+++ src/conf/parse.rs
@@ -55,7 +55,6 @@ fn opt_uc_x_help_message(_program: &str) -> String {
     let z_opts = concat!(
         "Options:\n",
         "  -X rust-version-info     display rust version info and exit\n",
-        "  -X base_dir=<path>       set <path> is base directory\n",
         "  -X map-ascii-rust-src    output rust source of the ascii map statistics\n",
     );
     z_opts.to_string()