+++ tests/helper.rs
@@ -41,6 +41,22 @@ macro_rules! help_msg {
 }
 
 #[allow(unused_macros)]
+macro_rules! x_help_msg {
+    () => {
+        concat!(
+            indoc::indoc!(
+                r#"
+            Options:
+              -X rust-version-info     display rust version info and exit
+              -X map-ascii-rust-src    output rust source of the ascii map statistics
+            "#
+            ),
+            "\n",
+        )
+    };
+}
+
+#[allow(unused_macros)]
 macro_rules! try_help_msg {
     () => {
         "Try --help for help.\n"